partly "reverting" the last commit in a sense that not sub elements are considered special but attributes. hene, i replaced the new tag @subelement with @visprop. @visprop is used to determine a visual property that is given via the attributes object on creation or via setProperty.
began unification of our nomenclature: we are often using a lot of different names for the same thing like "visprop", "property" and "attribute" for attributes like strokeColor and visible. Or "object" and "element" for geometry elements. We should agree on one term for every "thing" there is in jsxgraph to avoid confusion.

 so as a first consequence {g,s}etProperty was renamed to {g,s}etAttribute. getProperty is still available. as an alternative we could stick with {g,s}etProperty and rename the attributes parameter of the JXG.create* methods.