implements #3
This introduces extends the snapToGrid property to all elements. It is only relevant for points, lines and circles. For points, nothing changes, if this property is set for lines and circles they snap their parent points which have snapToGrid set to true, too, to the grid after the user finishes dragging this element.