fixes #2 If no element is dragged or found, don't stop the event propagation.
This is important on touch devices where the user might zoom completely onto the board and then is trapped because he can't change the viewport of the website anymore. Now buttons and links placed over the board should be reachable by the user (if the link or button is added via jsxgraph html texts, make sure the text has the property fixed set to true, otherwise if the link is too small, jsxgraph thinks the user wants to drag the text and prevents the event reaching the button or link).