new Makefile fixed some docstrings
The Makefile now uses node-jsdoc2 instead of the original jsdoc-toolkit. Not only does this speed up the docs generation but it also means it is very easy to install almost all dependencies which are required to build jsxgraph via a simple

$ npm install

The only dependency that has to be installed manually is nodejs.