Bugfix show/hide polygon
The behaviour of polygons was changed in 8e95f2c4. This is fixed by adding a parameter to JXG.Polygon.{show,hide}Element() that determines if the visibility state of borders of the polygon should be changed, too. Default is as it was in 0.96 and before: hide and show borders together with the polygon area.