Extend board.select() to allow a multi-element selection #29
This first implementation allows selecting elements based on attributes and properties of type number, string or boolean. A selector object can look like this:

{
  strokeColor: 'red',
  fixed: true,
  isDraggable: false
}

This selector will only select elements with the attribute set to red (it will *not* select elements with strokecolor set to #ff0000 or #f00), attribute fixed set to true and the isDraggable property of the element has to be set to false (this is usually set internally e.g. a point or a line with constraints).

Attribute lookup is currently case insensitive, properties lookup is case sensitive.