Implement new attributes for grid (#41)
The new grid attributes bottomRight and topLeft are given as arrays and restrict the area the grid is drawn in:

    board.create(grid, [], {bottomRight: [5, -5], topLeft: [-5, 5]});

This grid will only be displayed in the square defined by the two points. The coordinates will be interpreted as user coordinates, to use screen coordinates two other attributes are added: tlType and brType. Setting those to JXG.COORDS_BY_SCREEN will tell the grid to interpret the coordinates given in bottomRight and topLeft as screen coordinates.