Board selection mode
Allow a graph to be in a selectionMode by calling the function board.startSelectionMode().
Once in this mode, the mouse allow to draw a rectangle selection on the board.
Ideally, other mouse action, such as moving a point should be disabled.
To stop selectingMode, the function board.stopSelectionMode() must be called.
At any time, board.selectingBox gives mouse coordinates of the selection.
This commit doesn't include the rendering part.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis-lamarre@gmail.com>