usr2screen and rounding
This commit reverts / improves commit
22657cc8418676adaf8065236992eff2acaa943c.
The history of rounding screen coordinates in usrscreen() is:
In the beginning, we did rounding in this place, because old IE versions
could handle only integer values of screen coordinates in VML.

As far as I see, this rounding has been put into VML renderer since many
years and is not longer needed for stability.
Thus, we only round in usr2screen() if it is explicitly required.