Minor bug fix in initialisation of curve plotting
The _lastCrds variable in curve was being set to [NaN,NaN] rather than
the initially added point. This meant that if the first point evaluated
was a discontinuity the value of [NaN,NaN] would not be added to the
list of points as this appeared to be the last point (and as such would
be omitted to avoid repitition).