Added the 'msector' element.
The 'msector' (or 'm-sector') element, generalizes the bisector.
Instead of dividing an angle into two equal angles, it divides
the angle into two angles of amplitude m and (1-m).

The bisector is the special case in which m = 1/2.
The trisectors of an angle are the special cases in which m = 1/3 and m = 2/3.

This element is interesting because without it there is no way to conveniently
build trisectors, or other general divisions of an angle.
The ability to describe trisectors allows us to illustrate Morley's theorem.

There are two new examples:
- An example of a simple angle trisector
- An example illustrating morley's theorem