Regression fix: visibility
Visibility of tape measure points was wrong with commit
bf6fe235714822992487ac3d2ca9d7181bd12ff1 (ug fix: handling of inherit
visibility of points).
The problem was the position where visPropOld.visible is updated.
This is moved now to to the display-method of the renderer.
This fix made it necessary that the infobox is more treated as a regular
text element.