Polygon.removePoints: last-point removal bug fix
When checking whether the first or last point of the polygon had been
removed, so the border between them needed recreated, the index
considered to be "last" was hard-coded to 5, instead of computed (as
vertices.length-1). This caused incorrect behaviour when the polygon did
not have 5 vertices, and the last vertex (or vertices[5]) was removed.