Add examples for JSXGraph & C++ with WASM
We add a new example that shows how JSXGraph can be programmed with the
C++ programming language and how a JSXGraph construction in JavaScript
can utilize C++ code.

Instructions on how to compile the C++ sources to WASM can be found in
the README.md file in this commit. The compiled WASM module and runtime
JS script is added so the examples can be run without compiling it
first.