hideElement / showElement: not longer deprecated.
The deprecation for these two methods is removed.
Further, there are new shortcut methods: element.hide() and element.show().
All these methods are not longer declared private.

The reason is, there seem to be a demand for these methods.
element.setdisplayRendNode() does not meet this demand, it is private
method which is called internally and does not overwrite the visibility
attribute.

Fixes #345