JessieCode: if no attributes given to an object, its name is the LHS of the assignment
It looks like the intention was that when you create an object without
any explicit attributes, like:

    A = point(0,0):

the resulting point is given the name 'A'.

The bit of the 'op_assign' handler that store the name on the LHS was
broken - maybe the return value of getLHS changed since it was written?

I think it would be even more convenient if the LHS name was used even
if there are attributes given, but not the `name` or `id` attributes,
but for now this is a conservative fix to what the apparent intended
behaviour was.