Improve keyboard handling
Fixes #437

Panning of the construction is done with arrow keys
if the pan key (shift or ctrl - depending on the board attributes) is pressed.

Zooming is triggered with the keys +, o, -, if
the pan key (shift or ctrl - depending on the board attributes) is pressed.

Keyboard control is disabled if an HTML element of type input or textarea has received focus.