fix slider bug
This small change causes the label at the slider point to no longer be displayed in sketchometry. The error is probably caused by the change in `copyAttributes`. Why it does not appear in normal JSXGraph-applets is a mystery to me.

Nevertheless, this fix makes sense: `attr` is created by `copyAttributes` and therefore all contained attributes are in lower case anyway (see e.g. initialization of `withText`).