Use `push` to concatenate single elements
To concatenate an array that clearly has only one element, use `push`
directly, instead of indirectly through `Type.concat`. In Sphere3D, this
is needed to fix #641, because `Type.concat` doesn't handle singleton
arrays in the same way that `concat` does. In other elements, it's not
strictly needed, but I think it streamlines the code.