Bug fix: tangents on sectors
Fixes #685
Arcs and sectors are approximated by cubic Bezier curves.
Up to now, an arc was approximated by several Bezier curves spanning an
angle of pi/2 and the remaining angle.
Now, an arc is spanned by 4 Bezier curves spanning the same angle.
Otherwise, the position property of a glider can not be used for
determining the position of the glider in the Bezier path.