grid.getParents returns `parentAxes`
It was inheriting getParents from `Curve`, which led to an error when
trying to use that list to create a new grid object.

This overrides `getParents` on grid objects so that it returns just the
`parentAxes` list.

I think that dumping grids is still a bit wrong because the dump routine
produces separate objects for the major and minor grids, but at least it
doesn't throw an error any more.