Bug fix: color of arrow heads in Safari
Fixes #772
Since v1.12.1, the SVG value 'context-stroke' has been used to set the color
of arrow heads. The switch was necessary to support colors which are set with CSS, too.
As of Dec 2025, 'context-stroke' is not supported by Safari.
We updated JXG.isWebKitApple() for this and reactivated the
old setting of colors for Safari.